fix compile error, move test to existing class
